+++ packages/eyes-sdk-core/lib/frames/Frame.js
@@ -25,7 +25,6 @@ const EyesUtils = require('../EyesUtils')
  * @typedef {Object} SpecsFrame
  * @property {(selector) => boolean} isSelector - return true if the value is a valid selector, false otherwise
  * @property {(element) => boolean} isCompatibleElement - return true if the value is an element, false otherwise
- * @property {(leftElement: SupportedElement|EyesWrappedElement, leftElement: SupportedElement|EyesWrappedElement) => Promise<boolean>} isEqualElements - return true if elements are equal, false otherwise
  * @property {(logger: Logger, driver: EyesWrappedDriver, element: SupportedElement, selector: SupportedSelector) => EyesWrappedElement} createElement - return wrapped element instance
  */
 
@@ -125,17 +124,6 @@ class Frame {
     )
   }
   /**
-   * Equality check for two frame objects or frame elements
-   * @param {Frame|EyesWrappedDriver} leftFrame - frame object or frame element
-   * @param {Frame|EyesWrappedDriver} rightFrame - frame object or frame element
-   * @return {Promise<boolean>} true if frames are described the same frame element, otherwise false
-   */
-  static async equals(leftFrame, rightFrame) {
-    const leftElement = leftFrame instanceof Frame ? leftFrame.element : leftFrame
-    const rightElement = rightFrame instanceof Frame ? rightFrame.element : rightFrame
-    return this.specs.isEqualElements(leftElement, rightElement)
-  }
-  /**
    * @return {EyesWrappedElement}
    */
   get element() {
@@ -188,11 +176,12 @@ class Frame {
   }
   /**
    * Equality check for two frame objects or frame elements
-   * @param {Frame|EyesWrappedDriver} otherFrame - frame object or frame element
+   * @param {Frame|EyesWrappedDriver|UnwrappedElement} otherFrame - frame object or frame element
    * @return {Promise<boolean>} true if frames are described the same frame element, otherwise false
    */
   async equals(otherFrame) {
-    return this.constructor.equals(this, otherFrame)
+    const otherElement = otherFrame instanceof Frame ? otherFrame.element : otherFrame
+    return this._element ? this._element.equals(otherElement) : false
   }
   /**
    * Initialize frame reference by finding frame element and calculate metrics