+++ packages/eyes-webdriverio-5/src/SpecWrappedDriver.js
@@ -1,35 +1,100 @@
-const WDIOFrame = require('./WDIOFrame')
-const WDIOWrappedElement = require('./WDIOWrappedElement')
-const LegacySelector = require('./LegacySelector')
-const {remote} = require('webdriverio')
+const {TypeUtils} = require('@applitools/eyes-sdk-core')
 const {URL} = require('url')
+const {remote} = require('webdriverio')
+const LegacySelector = require('./LegacySelector')
+
+/**
+ * Supported selector type
+ * @typedef {string|Function|LegacySelector} SupportedSelector
+ */
+
+/**
+ * Compatible element type
+ * @typedef {UnwrappedElement|ResponseElement} SupportedElement
+ */
+
+/**
+ * Unwrapped element supported by framework
+ * @typedef {Object} UnwrappedElement
+ * @property {string} ELEMENT - legacy element id
+ * @property {string} element-6066-11e4-a52e-4f735466cecf - element id
+ */
+
+/**
+ * Response element the object returned from find element operation
+ * @typedef {Object} ResponseElement
+ * @property {UnwrappedElement} value
+ * @property {string} [selector]
+ */
+
+const LEGACY_ELEMENT_ID = 'ELEMENT'
+const ELEMENT_ID = 'element-6066-11e4-a52e-4f735466cecf'
 
 module.exports = {
-  isEqualFrames(leftFrame, rightFrame) {
-    return WDIOFrame.equals(leftFrame, rightFrame)
+  /* -------- ELEMENT -------- */
+
+  isCompatible(element) {
+    if (!element) return false
+    return Boolean(element.elementId || element[ELEMENT_ID] || element[LEGACY_ELEMENT_ID])
   },
-  createElement(logger, driver, element, selector) {
-    return new WDIOWrappedElement(logger, driver, element, selector)
-  },
-  createFrameReference(reference) {
-    return WDIOFrame.fromReference(reference)
+  isSelector(selector) {
+    return (
+      TypeUtils.isString(selector) ||
+      TypeUtils.isFunction(selector) ||
+      selector instanceof LegacySelector
+    )
   },
   toSupportedSelector(selector) {
-    return WDIOWrappedElement.toSupportedSelector(selector)
+    if (TypeUtils.has(selector, ['type', 'selector'])) {
+      if (selector.type === 'css') return `css selector:${selector.selector}`
+      else if (selector.type === 'xpath') return `xpath:${selector.selector}`
+    }
+    return selector
   },
   toEyesSelector(selector) {
-    return WDIOWrappedElement.toEyesSelector(selector)
+    if (selector instanceof LegacySelector) {
+      const {using, value} = selector
+      if (using === 'css selector') return {type: 'css', selector: value}
+      else if (using === 'xpath') return {type: 'xpath', selector: value}
+    } else if (TypeUtils.isString(selector)) {
+      const match = selector.match(/(css selector|xpath):(.+)/)
+      if (match) {
+        const [_, using, value] = match
+        if (using === 'css selector') return {type: 'css', selector: value}
+        else if (using === 'xpath') return {type: 'xpath', selector: value}
+      }
+    }
+    return {selector}
   },
+  extractId(element) {
+    return element.elementId || element[ELEMENT_ID] || element[LEGACY_ELEMENT_ID]
+  },
+  extractElement(element) {
+    return {
+      [ELEMENT_ID]: this.extractId(element),
+      [LEGACY_ELEMENT_ID]: this.extractId(element),
+    }
+  },
+  extractSelector(element) {
+    return element.selector
+  },
+  isStaleElementReferenceResult(result) {
+    if (!result) return false
+    return result instanceof Error && result.name === 'stale element reference'
+  },
+
+  /* -------- DRIVER -------- */
+
   async executeScript(driver, script, ...args) {
     return driver.execute(script, ...args)
   },
-  sleep(driver, ms) {
+  async sleep(driver, ms) {
     return driver.pause(ms)
   },
-  switchToFrame(driver, reference) {
+  async switchToFrame(driver, reference) {
     return driver.switchToFrame(reference)
   },
-  switchToParentFrame(driver) {
+  async switchToParentFrame(driver) {
     return driver.switchToParentFrame()
   },
   async findElement(driver, selector) {
@@ -45,22 +110,28 @@ module.exports = {
     return Array.from(elements)
   },
   async getWindowLocation(driver) {
-    const rect = await driver.getWindowRect()
-    return {x: rect.x, y: rect.y}
+    const location = TypeUtils.isFunction(driver.getWindowPosition)
+      ? await driver.getWindowPosition()
+      : await driver.getWindowRect()
+    return {x: location.x, y: location.y}
   },
-  async setWindowLocation(driver, location) {
-    return driver.setWindowRect(location.x, location.y, null, null)
+  async setWindowLocation(driver, location = {}) {
+    // devtools protocol doesn't support change position on runtime
+    if (driver.isDevTools) return
+    return TypeUtils.isFunction(driver.setWindowPosition)
+      ? driver.setWindowPosition(location.x, location.y)
+      : driver.setWindowRect(location.x, location.y, null, null)
   },
   async getWindowSize(driver) {
-    const rect = driver.hasOwnProperty('getWindowRect')
-      ? await driver.getWindowRect()
-      : await driver.getWindowSize()
-    return {width: rect.width, height: rect.height}
+    const size = TypeUtils.isFunction(driver.getWindowSize)
+      ? await driver.getWindowSize()
+      : await driver.getWindowRect()
+    return {width: size.width, height: size.height}
   },
-  async setWindowSize(driver, {x = null, y = null, width, height} = {}) {
-    return driver.hasOwnProperty('setWindowRect')
-      ? driver.setWindowRect(x, y, width, height)
-      : driver.setWindowSize(width, height)
+  async setWindowSize(driver, size = {}) {
+    return TypeUtils.isFunction(driver.setWindowSize)
+      ? driver.setWindowSize(size.width, size.height)
+      : driver.setWindowRect(null, null, size.width, size.height)
   },
   async getOrientation(driver) {
     const orientation = await driver.getOrientation()
@@ -103,48 +174,64 @@ module.exports = {
     return driver.url(url)
   },
 
-  /********* for testing purposes */
+  /* -------- TESTING -------- */
 
-  async build({capabilities, serverUrl = process.env.CVG_TESTS_REMOTE}) {
-    const {hostname, port, pathname, protocol} = serverUrl ? new URL(serverUrl) : {}
-    let fixedPort = port
-    if (protocol === 'http:' && !port) {
-      fixedPort = 80
-    }
-    if (protocol === 'https:' && !port) {
-      fixedPort = 443
-    }
-    const options = {
-      logLevel: 'silent',
-      capabilities: capabilities,
-      path: pathname,
-      port: fixedPort ? Number(fixedPort) : undefined,
-      hostname,
-      protocol: protocol ? protocol.replace(/:$/, '') : undefined,
+  async build({capabilities, server, logLevel = 'silent', protocol = 'webdriver'} = {}) {
+    const options = {capabilities, logLevel, automationProtocol: protocol}
+    if (server && server.url) {
+      const url = new URL(server.url)
+      options.protocol = url.protocol ? url.protocol.replace(/:$/, '') : undefined
+      options.hostname = url.hostname
+      options.port = Number(url.port)
+      options.path = url.pathname
+      if (server.url.includes('saucelabs.com')) {
+        if (server.w3c) {
+          if (!options.capabilities['sauce:options']) {
+            options.capabilities['sauce:options'] = {}
+          }
+          options.capabilities['sauce:options'].username = server.username
+          options.capabilities['sauce:options'].accessKey = server.accessKey
+        } else {
+          options.capabilities.username = server.username
+          options.capabilities.accessKey = server.accessKey
+        }
+      } else if (server.url.includes('browserstack.com')) {
+        if (server.w3c) {
+          if (!options.capabilities['bstack:options']) {
+            options.capabilities['bstack:options'] = {}
+          }
+          options.capabilities['bstack:options'].userName = server.username
+          options.capabilities['bstack:options'].accessKey = server.accessKey
+        } else {
+          options.capabilities['browserstack.user'] = server.username
+          options.capabilities['browserstack.key'] = server.accessKey
+        }
+      }
     }
     return remote(options)
   },
-
   async cleanup(driver) {
     return driver.deleteSession()
   },
-
   async click(driver, element) {
     const extended = await driver.$(element)
     return extended.click()
   },
-
   async type(driver, element, keys) {
     const extended = await driver.$(element)
     return extended.setValue(keys)
   },
-
   async waitUntilDisplayed(driver, selector, timeout) {
     const el = await this.findElement(driver, selector)
     return el.waitForDisplayed({timeout})
   },
-
-  async getElementRect(driver, el) {
-    return driver.getElementRect(el.elementId)
+  async getElementRect(_driver, element) {
+    if (TypeUtils.isFunction(element.getRect)) {
+      return element.getRect()
+    } else {
+      const size = await element.getSize()
+      const location = await element.getLocation()
+      return {...location, ...size}
+    }
   },
 }